Fix 404 when visiting /resume
I've actually removed the redirect to /wal_resume.pdf from /resume. Since
the resume is handled by the server and routing is client side, it doens't
work (404). The same thing happened when I use the React Router <Link/>
component to navigate to the resume, so I added back the "external" link
option for the PageButton component in the navBar menu.