Apply Michael's suggestions
Last time I tried to fix the W svg in the portfolip project page, I couldn't
get the round line-cap to work. It worked this time, and it fixes the gap
created when line makes its way back to the start.

He also pointed out a typo in my Mobile Dev description on my homepage.